fix: 安全更新的功能缺失
处理数据传输问题，原逻辑修改

Log: 安全更新的功能缺失
Bug: https://pms.uniontech.com/bug-view-173629.html
Influence: 控制中心-更新-安全更新

+++ src/plugin-update/operation/updatedbusproxy.h
@@ -50,8 +50,8 @@ public:
     bool autoClean();
 
     Q_PROPERTY(qulonglong UpdateMode READ updateMode WRITE setUpdateMode NOTIFY UpdateModeChanged)
-    qulonglong updateMode();
-    void setUpdateMode(qulonglong value);
+    uint updateMode();
+    void setUpdateMode(uint value);
 
     QList<QDBusObjectPath> jobList();
     QDBusPendingReply<QDBusObjectPath> UpdateSource();
